fixed database

+++ App/models/livestock.py
@@ -4,7 +4,7 @@ class Livestock(db.Model):
   id = db.Column(db.Integer, primary_key=True)
   name =  db.Column(db.String(255), nullable=False)
   quantity = db.Column(db.Integer)
-  price = db.Column(db.Double)
+  price = db.Column(db.Integer)
   noOfAdults = db.Column(db.Integer)
   noOfYoung = db.Column(db.Integer)
